requirements.txt
flask
flask-socketio
paho-mqtt
google-api-python-client
google-auth-oauthlib
gunicorn
eventlet
python-dotenv
